+++ docs/models.md
@@ -1,3 +1,3 @@
-::: qadence.quantum_model
+::: qadence.model
 
 ::: qadence.ml_tools.qnn
