+++ tests/models/test_quantum_model.py
@@ -15,9 +15,9 @@ from qadence.circuit import QuantumCircuit
 from qadence.constructors import hea, total_magnetization
 from qadence.divergences import js_divergence
 from qadence.ml_tools.utils import rand_featureparameters
+from qadence.model import QuantumModel
 from qadence.operations import MCRX, RX, HamEvo, I, Toffoli, X, Z
 from qadence.parameters import FeatureParameter, VariationalParameter
-from qadence.quantum_model import QuantumModel
 from qadence.states import equivalent_state
 from qadence.transpile import invert_endianness
 from qadence.types import PI, BackendName, DiffMode
