+++ tests/qadence/test_measurements/test_shadows.py
@@ -26,9 +26,9 @@ from qadence.measurements.shadow import (
     local_shadow,
     number_of_samples,
 )
+from qadence.model import QuantumModel
 from qadence.operations import RX, RY, H, I, X, Y, Z
 from qadence.parameters import Parameter
-from qadence.quantum_model import QuantumModel
 from qadence.serialization import deserialize
 from qadence.types import BackendName, DiffMode
